nueva clase

+++ clase 4/Ejercicios.java
@@ -1,18 +1,21 @@
-public class Ejercicios{
+public class Ejercicios {
     public static void main(String[] args) {
-        int angulo = 80;
-        if(angulo == 90){
+        double angulo = 80;
+        if (angulo == 90) {
             System.out.println("es un angulo recto");
-        }else {
+        } else {
             System.out.println("no es un angulo recto");
         }
-    }    
-    //------------------------------------------
-        int temp = 101;
-        if(temp > 100){
-            System.out.println("por encima del punto de ebullicion del agua");
-        }else{
-            System.out.println("por debajo del punto de ebullcion del agua");}
-// ------------------------------------------
+    }
+
+    // ------------------------------------------
+    int temp = 101;if(temp>100)
+    {
+        System.out.println("por encima del punto de ebullicion del agua");
+    }else
+    {
+        System.out.println("por debajo del punto de ebullcion del agua");
+    }
+    // ------------------------------------------
 
 }
